xbill/requirements.txt
Django==1.5.1
wsgiref==0.1.2
pycrypto==2.6.1
django-widget-tweaks==1.3
python-social-auth==0.1.14
psycopg2==2.5.1
python-dateutil==2.1
South==0.8.2
django-compressor==1.3
django-braces==1.2.2
django-extra-views==0.6.4
pika==0.9.13
djangorestframework==2.3.13
Markdown==2.4
django-filter==0.7
django-pdb==0.3.2
boto==2.28.0